refactor: add interface for OpenIdConnectClient
